Add function randomizeSeating
This function responds to a message action on a slack message.
The bot will randomize the list of users that reacted with a valid
reaction, and reply to the message with that list. If there is no valid
reactions the bot will send an ephemeral message to the user invoking
the message action.